requirements.txt
Pillow
pytesseract
pdf2image